requirements.txt
fastapi == 0.111.0
uvicorn == 0.30.1
starlette == 0.37.2
uvicorn == 0.30.1
sqlalchemy == 2.0.31
pydantic == 2.8.2
FastAPI-SQLAlchemy  == 0.2.1
python-dotenv == 1.0.1
pytest == 8.2.2
psycopg2 == 2.9.9
alembic == 1.13.2
python-multipart == 0.0.12
httpx == 0.27.2
psycopg2-binary == 2.9.10
passlib == 1.7.4
pika == 1.3.2
python-jose == 3.3.0
bcrypt == 4.2.0
image == 1.5.33
replit == 4.1.0